refactor: deleted unused code refactor: changed namespaces chore: update LightResults to 8.0.11
