+++ src/McdaToolkit/Mcda/Factories/Enums/McdaMethods.cs
@@ -1,0 +1,7 @@
+﻿namespace McdaToolkit.Mcda.Factories.Enums;
+
+public enum McdaMethods
+{
+    TOPSIS = 0,
+    VIKOR = 1
+}
