+++ src/McdaToolkit/Mcda/Services/MatrixChecker/Errors/DecisionCriteriaHaveIncorrectValueError.cs
@@ -1,5 +1,5 @@
 ﻿using LightResults;
 
-namespace McdaToolkit.Mcda.Helpers.Errors;
+namespace McdaToolkit.Mcda.Services.MatrixChecker.Errors;
 
 public class DecisionCriteriaHaveIncorrectValueError() : Error("Criteria decision types should be number ∈Z{-1;1}");
