+++ src/McdaToolkit/Mcda/Services/MatrixChecker/Errors/MatrixColumnLengthNotEqualWeightsVectorLengthError.cs
@@ -1,5 +1,5 @@
 ﻿using LightResults;
 
-namespace McdaToolkit.Mcda.Helpers.Errors;
+namespace McdaToolkit.Mcda.Services.MatrixChecker.Errors;
 
 public class MatrixColumnLengthNotEqualWeightsVectorLengthError() : Error("Columns length of data matrix should be equal length of weights and types arrays");
