+++ src/McdaToolkit/Mcda/Services/MatrixChecker/Errors/WeightNotSumToOneError.cs
@@ -1,5 +1,5 @@
 ﻿using LightResults;
 
-namespace McdaToolkit.Mcda.Helpers.Errors;
+namespace McdaToolkit.Mcda.Services.MatrixChecker.Errors;
 
 public class WeightNotSumToOneError() : Error("Sum of weight have to equal 1");
