+++ src/McdaToolkit/McdaToolkit.csproj
@@ -28,7 +28,7 @@
     </ItemGroup>
 
     <ItemGroup Label="Dependencies">
-        <PackageReference Include="LightResults" Version="8.0.10" />
+        <PackageReference Include="LightResults" Version="8.0.11" />
         <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
     </ItemGroup>
 </Project>
