+++ src/McdaToolkit/Normalization/Services/MatrixNormalizator/Errors/NormalizationMethodsEqualError.cs
@@ -1,5 +1,5 @@
 ﻿using LightResults;
 
-namespace McdaToolkit.Normalization.Service;
+namespace McdaToolkit.Normalization.Services.MatrixNormalizator.Errors;
 
 public class NormalizationMethodsEqualError() : Error("New method have to be diffrent than current one");
